+++ osu.Game.Rulesets.Karaoke/UI/HUD/BindableNumberExtension.cs
@@ -1,0 +1,23 @@
+﻿// Copyright (c) andy840119 <andy840119@gmail.com>. Licensed under the GPL Licence.
+// See the LICENCE file in the repository root for full licence text.
+
+using osu.Framework.Bindables;
+
+namespace osu.Game.Rulesets.Karaoke.UI.HUD
+{
+    /// <summary>
+    /// Will move into framework layer
+    /// </summary>
+    public static class BindableNumberExtension
+    {
+        public static void TriggerIncrease(this BindableInt bindableInt)
+        {
+            bindableInt.Value += bindableInt.Precision;
+        }
+
+        public static void TriggerDecrease(this BindableInt bindableInt)
+        {
+            bindableInt.Value -= bindableInt.Precision;
+        }
+    }
+}
