+++ osu.Game.Rulesets.Karaoke/UI/HUD/SettingOverlayContainer.cs
@@ -4,7 +4,6 @@
 using System;
 using System.Linq;
 using osu.Framework.Allocation;
-using osu.Framework.Bindables;
 using osu.Framework.Graphics.Containers;
 using osu.Framework.Input.Bindings;
 using osu.Framework.Input.Events;
@@ -76,22 +75,6 @@ namespace osu.Game.Rulesets.Karaoke.UI.HUD
                 dependencies.CacheAs(drawableKaraokeRuleset.Session);
 
             return dependencies;
-        }
-    }
-
-    /// <summary>
-    /// Will move into framework layer
-    /// </summary>
-    public static class BindableNumberExtension
-    {
-        public static void TriggerIncrease(this BindableInt bindableInt)
-        {
-            bindableInt.Value += bindableInt.Precision;
-        }
-
-        public static void TriggerDecrease(this BindableInt bindableInt)
-        {
-            bindableInt.Value -= bindableInt.Precision;
         }
     }
 }
